changes, requires fix to compilation errors

+++ tag_engine/src/MeshClass.cpp
@@ -317,16 +317,16 @@ void TAGMesh::draw(const TAGShaderManager::Shader& shader, const TAGShaderManage
 			shader.set<glm::vec3>(options.colour_vec, material.colour);
 		}
 		else {
-			shader.set<unsigned int>(options.diffuse_tex_array, diffuse[0], diffuse.size());
+			shader.set<unsigned int>(options.diffuse_tex_array, diffuse[0], (unsigned int) diffuse.size());
 		}
-		shader.set<unsigned int>(options.diffuse_tex_num, diffuse.size());
+		shader.set<unsigned int>(options.diffuse_tex_num, (unsigned int) diffuse.size());
 
 		if (material.spec_fac > 0.0f && specular.size() > 0) {
-			shader.set<unsigned int>(options.specular_tex_array, specular[0], specular.size());
+			shader.set<unsigned int>(options.specular_tex_array, specular[0], (unsigned int) specular.size());
 		}
 		shader.set<float>(options.specular_factor, material.spec_fac);
 		shader.set<float>(options.specular_exp, material.spec_exp);
-		shader.set<unsigned int>(options.specular_tex_num, specular.size());
+		shader.set<unsigned int>(options.specular_tex_num, (unsigned int) specular.size());
 
 		shader.set<float>(options.opacity_value, material.opacity);
 		shader.set<glm::vec3>(options.camera_pos, TAGBaseState::camera_position);
